requirements.txt
kafka-python
spark
pyspark
cassandra-driver